Rewrite loaders to return just the new page
Returning the entire feed on every 'more' meant mutations to earlier items in the feed would get overwritten without using Object.assign() or something to directly update the original view. This makes the generator functions return just the new page so it can be concatenated and the old data stays unchanged.