Use the Virtual Feed with comments
Some comment feeds are also really long and hard to render, this required rewriting the CommentTree to generate a flat structure instead of a nested one, so nested comments are counted too when rendering by an index range. They've been styled differently too in a way that fits better with this layout. The VirtualFeed still has some jankiness that needs to be solved, but this at least gets everything onto the same page.

The "Load More" buttons now also will disable and show a spinner while waiting for more comments to load.