Optimize the virtual feed
Instead of using intersection observers before and after the items to detect scroll direction and shifting the index +/-1 with each detection, it now always uses the handler for when scrolling the scroll container (the feed now looks for a specific class to find a scroll container for comments on the post overlay where the document itself isn't the one scrolling).

The virtual feed component is a little better now. It's still not perfect (rendering what the viewport can fit instead of some arbitrary number of items would be nice) but this should eliminate a few bugs like scrolling to the bottom of the feed before things can render in now won't leave you sitting on the "Load more" banner even when it should be showing stuff.