Add a maxlength to the password field to match lemmy-ui
On the official sites, there's a 60 character maxlength on password fields, so if a user were to sign up with a >60 character password they'd have no idea it just gets truncated beyond that, but with no limit on Alexandrite they'd submit their full password and it's suddenly saying the password is wrong. This adds a limit to match what lemmy-ui is doing.