Calculate rendered items based on what actually fits + overscan
Instead of just using some hard coded min/max values, this actually goes the distance and calculates how many are required to fill the screen (plus a bit extra on the top/bottom so the user can scroll without seeing a blank feed if we haven't re-calculated yet).

This is much more flexible, and able to handle shorter items in a feed without having to manually raise the max which would either be too little or overkill depending on screen height and the sizes of things.