Added comment searching
This adds a search bar to comments, letting you filter comments by some text. This is nice because you can't really ctrl+f effectively with the virtual feed. Threads with no matches inside won't show, and any non-match parent of a search match shows but won't show the comment content until you hover over it, so you can see the context for all search matches.

This was the primary motivation for the VirtualFeed rendering rewrite, because on a 1440p screen I was hitting the default max rendered items with what was visible on screen. Now it dynamically increases to fit regardless.