Prevent duplicate comments from showing up if added while viewing a single comment thread
It seeds the PostPage comments with comments from a load function unlike the other places. The comment submit form causes the load to run and the comment is added both from running load and the event dispatched by the comment form which usually does it.

I made it only use the load's comments at first, and ignore it after so it behaves like all the other pages.