Properly use the default value if a setting's previous value isn't found
I got it wrong and it doesn't throw when JSON.parse tries to parse null, so `null` was used instead of the fallback. I was banking on that throwing so if there was no previous value found in localStorage it'd use the default instead of overwriting with some blank value but I was wrong, now it'll check for null first and only try and overwrite with the JSON parsed value if there was something there.