Make commenting/editing/replying and voting work client-side
I also rewrote how many async actions are tracked. There are two wrapper functions for either form submit or just a button click handler for tracking pending/error states as well as wrapping the submit function conveniently. This makes it a lot less messy to track state/progress of things, and will make it a lot cleaner when I add more features in the future.