Detect invalid sessions and redirect to the login screen
If we've passed an auth token, but don't get a local user back when the page loads we can assume the user got logged out, as there doesn't seem to be another good check I've found so far (invalid auth tokens on getPosts calls don't seem to return an error). Additionally the fetch function should throw redirects in the future, but this is a good first step.