Add some customization settings, more timely inbox unread checking
There are now a couple of settings, the theme of the website and how to handle NSFW stuff. I created a framework for using stores for Alexandrite settings that save to local storage.

I also added some calls to update the header unread counter when I know it'd change, like when replying to an unread comment from the post it's on.

Images no longer retry loading, that was for when Lemmy had serious stability problems and image load chance of success was like 50%.