Added a new feed layout, rearranged Post component
Now instead of just having access to the overlay, you can view stuff side-by-side in columns, making it a little less tedious as you don't have to close the overlay in between viewing posts.

To fit things better in a narrower column I had to rearrange things. Now the "show in overlay" button is by the rest of the buttons, as you can now also click the post title to open it in the overlay. Clicking the post title now opens it in an overlay as long as that's supported by the page.

I also had to rearrange the feeds, and now virtual scrolling the feed or the column post won't scroll the body, just their immediate parents, it looks very similar but was necessary.