Normal layout viewing comments, fix viewing comments directly
I had added some code to reload comments based on the postView changing, so it can load new comments when viewing a different post in a feed's column view, but this had cause it to wipe out the initial comments we load on a `/comments/<commentId>` route. It doesn't do that now.