Add "Auto" feed layout, linting
There's now a default "AUTO" feed layout option. It tries to guess at how wide all the things that are fitting on screen are (feed + post + sidebar (optional) width estimates), and will show the overlay layout if the screen is too narrow, or the column layout if there's enough space. This includes the sidebar (and whether it's visible) into the calculations, so it can switch on the fly, and it won't squish the post which is the main flexible width element (if you use layout on a narrow screen with the sidebar open it makes the post illegibly narrow).

I also fixed a bug with the virtual feed, where a really tall element could take up the entire viewport and scrolling would get stuck. Now it takes into account how far past that the viewport is so it will not go strictly based on height of the rendered content, but ensures it renders all the way to the target `top` coordinate of the bottom of the rendered content.

I also fixed all the existing lint errors.