Replace duplicated feed data handling with a common store, VirtualFeed buffer store for comments
Basically every page that had a list of content on it had a lot of similar but duplicated code. Especially when it came to applying updates from a nested component. Most of the time when you'd do an action that caused an update it'd dispatch an `update-post-view` or similar event with the updated thing, but that meant that every page that held a list of posts needs to listen to that and use the updated post instead of the stale one it has in memory. This meant that if a page didn't have all that duplicated stuff, things like upvotes wouldn't work (like in the inbox). This now puts a store that handles all kinds of content in context, wrapping every feed, so nested components can easily update their content without having to play some game of telephone with making sure to add `on:update-post-view` to every intermediate component, and on every page that component is on.

I also added a new virtual feed "buffer" store, which is just a `Record<string, any>` that lets you store random temp data in context, so scrolling an item out of view doesn't cause you to lose data. Previously when you would start typing a reply, if you scrolled out of view then back the reply composer would be closed and it would be missing your text. Now it sticks around.

Also fixed a markdown bug where a sequence of spoilers would be nested, turns out the match only work on the start of the spoiler so it just wasn't ever closing the 'details', added a message to disable DarkReader in the theme settings.