Improved the markdown editor
The markdown editor I was using was just the comment composer component that I reused and shoehorned in anywhere you'd use formatted text. This has changed! Now it's its own component without all of the extra features, now every other usage of the editor will have to use multiple components to get the functionality it wants.

The editor also has been improved and includes buttons to add formatting to text by surrounding the selected text with various characters.

I used it to add the Bio to the settings.