Don't duplicate newly added comments
When adding a comment the comment gets prepended to the comment list, so it shows up at the top, making it easier to see that the user just added that comment, but if the next page loaded it'd also add it from the next page of comments. This fixes that.