Use thumbnails for user/community icons
I missed this! They were loading the full sized images, now they use much smaller ones so things should be more efficient, and maybe it's placebo but the scrolling feels smoother on the feed.

The alternate sources for images also now check if the URL looks like an image hosted by Lemmy (just checking that the pathname starts with `/pictrs/`, so it doesn't try using sources that are likely invalid if loading something from another server.