Add a preview of the parent comment when context is unclear
When you are deep in a comment thread and suddenly the thread ends and it goes back a couple levels you suddenly don't have any context in front of you as to what that comment is a reply to. So now when a comment is at a shallower level than the last it will show an icon you can hover to see a preview of the parent comment.