Account switcher, instance in path, bug fixes
This continues with the profile stuff and adds a nicer profile switcher. The other main thing is now instances are part of the route. You can also set a default account per instance (and a default instance if navigating to `/`).

Also fixed some linting errors, and some weirdness with sending messages.