Instance switcher improvements, siteMeta as a store, routing improvements
The user menu now lets you switch which instance you're viewing (in which default account rules apply), and there were some bugs fixed for the instance page crashing trying to render an avatar (no user so no settings context).

The siteMeta app context is now a store so it can be updated when switching users and things like the list of subscriptions in the sidebar get updated and it shows you mod action buttons on the correct communities.

The app now will properly redirect you from an instanceless route, if you have a path that looks like the start of a instance level route (u, c, post, etc) it'll redirect you with your default instance (or a fallback). This also meant I could remove the "guard" at the top of the layout load which didn't really work and was causing every page navigation to reload the siteMeta which is a big waste.