Make the community favorite button more easily discoverable
The button to favorite a community was only showing up if you hovered the icon. Very few people thought to do this so people didn't think this was a feature. Now if you hover over the entire community bar it'll make the icon mostly transparent and show the button so it should be more obvious to anyone who uses the sidebar.