Added/corrected language settings, community moderator badges
Now you can select your languages on the Lemmy settings tab, and those languages are used, along with the community discussion languages to filter to what is a valid language there.

Now that I need to load the community to get discussion languages for comments and posts I'm also using that to put a badge next to moderators in the comments section of a post.