Add some hotkeys, auto-focus post lists/posts, click anywhere on cards to open them
I've added a few hotkeys and a new Help page that lists them.

When loading the page the feed (or the post if viewing the post) will automatically be focuses so you can use page up/down arrows etc to scroll. Fixes #41

You can also now click anywhere on a card, not just the title/image, to open it.