Use the PostBody component when viewing posts, nicer mobile sizing
This change mostly impacts things when viewing posts. The three "To Comments/Hide Content/View Source" buttons were difficult to make look good on mobile, so the "Hide Content" button is gone (it was redundant with the same button in the post preview above it), and now the View Source button is part of the PostBody component, which is now used here instead of just putting the Markdown directly on the page.

The PostBody component also doesn't have the `" Post` 'header' anymore, as it was kind of useless. The new "View Source" just floats in the top right corner.

I also changed the sizing of a bunch of buttons so they look nicer on mobile.