Don't sync unread counts between tabs
Syncing between tabs is great, but if you're logged in as two different profiles in different tabs they'll fight over what the real count is, and if they're not the same they'll alternate whenever each page polls. For now I'm just reverting using a localStorage backed store for those and plan to re-add it when I make settings unique to each profile which requires a lot more effort.