When resigning as a mod, hide the button afterwards
Now checking the community context to see if the logged in user is a mod, instead of `$siteMeta`, as we only update the community context when that happens.

To cover all the bases it also will remove it from siteMeta if the user is resigning, so we hide all the other moderation buttons.