Remove the focus outline of the post feed background
Initial focus is placed on the post feed on page load for up/down/page up/page down/etc keys to work from the start. This however added an outline on that element as part of default user agent styling in Firefox for focused elements I think, and the way the background of the header looked it meant you'd see a white gradient from the bottom of the header which looked different depending on the computer but was unwanted everywhere.

This just removes that outline.