Add image uploading via markdown editor
This lets you paste (currently only that, need to add a file input for people who want to select the file) an image into a markdown editor and it'll upload the image. After uploading an image a "Manage Uploads" button appears in the top right of the component which opens a modal that shows you image URLs and lets you delete images. Deleting an image also will remove the markup for that image from the comment.