Add pure light/dark theme setting
This is for people who want a pure dark or pure white background instead of a variety of shades of dark and light gray with a hue tint. It applies a `.pure-theme` class to `<html>` and overrides styles on a bunch of types of container with backgrounds with just that theme's color. Some additional borders are given to those elements if there was none to keep it grouped together.